// src/types/bindings/CreditNoteUpdatePayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CreditNoteItemUpdatePayload } from "./CreditNoteItemUpdatePayload";

export type CreditNoteUpdatePayload = { credit_note_number: string, credit_note_date: string, remarks: string | null, reason: string | null, items: Array<CreditNoteItemUpdatePayload>, expected_revision_no: number, };